Reworked logger + basic rewards for combat + combat screen as pop up
Now, CombatLogger instance is created for each combat encounter, that way its gonna be possible for more than one combat encounter to be active. Added basic rewards for combat, which are awarded once the combat screen is closed. Also added combat screen which is going to have an option to view combat log.

+++ Assets/Scripts/Guild/GuildMain.cs
@@ -16,7 +16,7 @@ public class GuildMain : MonoBehaviour
     public delegate void GuildResourceChangedDelegate();
     public GuildResourceChangedDelegate GuildResourceChanged;
 
-    public int gold;
+    public float gold;
 
     private void Awake()
     {
@@ -26,7 +26,7 @@ public class GuildMain : MonoBehaviour
         }
     }
 
-    public bool SubtractGold(int amount)
+    public bool SubtractGold(float amount)
     {
         if(amount > gold)
         {
@@ -39,7 +39,7 @@ public class GuildMain : MonoBehaviour
             return true;
         }
     }
-    public bool AddGold(int amount)
+    public bool AddGold(float amount)
     {
         if (amount > gold)
         {
